+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Version 2.8.10 - TBD 
+  * Add legend for events on map (#192)
   * Add option to apply helicorder view settins to all open helicorders (#237)
   * Fixes for non-integer sample rate (#260)